Upload three openings books, based on Stephan Pohl's work.
The first book is UHO_4060_v2.pgn, which is a reverse engineered .pgn to Stockfish's UHO_4060_v2.epd.
The second book, UHO_4060_vT.pgn, reverse engineers the same set of data, which was filtered on Torch instead of SF.
The third book, UHO_4060_v2T.pgn, is the intersection of the two books. The default now for CCC23 and onwards